signal: log processes crashes
If the log_fatal_signals sysctl variable is set to non-zero value,
the kernel the following string to the kernel log:

"pid <pid>, uid <uid>: exited on signal <signal number> [(core dumped)]".

This behavior exactly repeats the FreeBSD behavior.

Note, that there is also print-fatal-signals sysctl. It's used
for debug purposes and cannot be used for Yandex monitoring.
We do not change it's behavior to support full compatibility.

Signed-off-by: Roman Gushchin <klamm@yandex-team.ru>